+++ test/sneaker_handlers/retry_handler_test.rb
@@ -4,31 +4,13 @@ require "sneakers/runner"
 require "rabbitmq/http/client"
 require 'pry'
 require 'json'
+require 'test_worker'
 
 class SneakersHandlers::AcceptanceTest < Minitest::Test
 
-  class TestWorker
-    include Sneakers::Worker
 
-    from_queue "sneaker_handlers.dead_letter_success_test",
-    ack: true,
-    durable: false,
-    exchange: "sneakers_handlers",
-    exchange_type: :topic,
-    routing_key: "sneakers_handlers.dead_letter_test",
-    handler: SneakersHandlers::RetryHandler,
-    arguments: { "x-dead-letter-exchange" => "sneakers_handlers.dlx",
-                 "x-dead-letter-routing-key" => "sneaker_handlers.dead_letter_success_test" }
 
-    def work(payload)
-      JSON.parse(payload)
-      response = payload["response"]
-      x = JSON.parse(payload)["response"] + "!"
-      return reject!
-    end
-  end
-
-  def test_dead_letter_messages
+  def test_max_retry_goes_to_dlx
     delete_test_queues!
     configure_sneakers
 
@@ -43,11 +25,9 @@ class SneakersHandlers::AcceptanceTest < Minitest::Test
       }
     JSON
 
-    1.times do
-      exchange.publish(json, routing_key: "sneakers_handlers.dead_letter_test")
-    end
+    exchange.publish(json, routing_key: "sneakers_handlers.dead_letter_test")
 
-    sleep 5
+    sleep 5 #wait for the worker to deal with messages
 
     dead_letter = channel.queue(TestWorker.queue_name + ".dlx")
     assert_equal 1, dead_letter.message_count
